requirements.txt
fFlask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
gunicorn
python-dotenv
